block-buffer: add LazyBlockBuffer and DigestBuffer trait
